+++ src/components/shared/PrintPretty.vue
@@ -1,24 +1,13 @@
 <template>
   <div>
-    <v-row>
-      <v-col class="my-0 py-0" cols="12" md="9">
-        <h2 v-if="caption" class="title">{{caption}}</h2>
-      </v-col>
-      <v-col class="my-0 py-0" cols="12" md="3">
-        <v-btn-toggle v-model="wrapLines" class="float-right v-btn-toggle--small">
-          <v-btn :value="true" title="Wrap lines">
-            <v-icon>mdi-wrap</v-icon>
-          </v-btn>
-        </v-btn-toggle>
-      </v-col>
-    </v-row>
+    <h2 v-if="caption" class="title">{{caption}}</h2>
     <div v-if="resizable" class="py-1">
       <resizable-container :initial-height="initialHeight">
         <code-editor :value="document" read-only/>
       </resizable-container>
     </div>
     <div v-else :style="style" class="pt-2">
-      <code-editor :value="document" :wrap-lines="wrapLines" read-only/>
+      <code-editor :value="document" read-only/>
     </div>
   </div>
 </template>
@@ -52,11 +41,6 @@
       caption: {
         type: String,
         default: ''
-      }
-    },
-    data () {
-      return {
-        wrapLines: true
       }
     },
     computed: {